review comments. renamed hash to hash_query and fixed fmt issues
